requirements.txt
# Основные зависимости
pydantic>=2.0.0
pyyaml>=6.0

# LLM
ollama>=0.1.0

# Форматирование кода (опционально)
black>=23.0.0
ruff>=0.1.0
autopep8>=2.0.0

# Работа с репозиториями
python-gitlab>=3.0.0
GitPython>=3.1.0

# CDP интеграция
playwright>=1.40.0

